backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.2
python-dotenv>=1.0
Pillow>=10.0
pandas>=2.2
openpyxl>=3.1
whitenoise>=6.7
# Optional for Postgres (uncomment if you switch):
# psycopg2-binary>=2.9
